backend_configured/requirements.txt

aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
# anyio==3.7.1
asttokens==3.0.0
attrs==25.3.0
cachetools==5.5.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.2.0
dataclasses-json==0.6.7
decorator==5.2.1
distro==1.9.0
executing==2.2.0
fastapi==0.115.1
filelock==3.18.0
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.3.2
google-ai-generativelanguage==0.6.18
google-api-core==2.24.2
google-auth==2.40.1
google-genai==1.21.1
googleapis-common-protos==1.70.0
greenlet==3.2.1
groq==0.26.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.30.2
idna==3.7
ipython==8.18.1
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-google-genai==2.1.4
langchain-groq==0.3.2
langchain-huggingface==0.1.2
langchain-tavily==0.2.10
langchain-text-splitters==0.3.8
langgraph==0.4.8
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.66
langgraph-swarm==0.0.11
langsmith==0.4.5
linkify-it-py==2.0.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
memory-profiler==0.61.0
memray==1.17.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.6.3
mypy_extensions==1.1.0
networkx==3.4.2
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
orjson==3.10.18
ormsgpack==1.10.0
packaging==23.2
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
pip==24.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pydantic-settings==2.10.1
Pygments==2.19.2
python-dotenv==1.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.1.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
setuptools==65.5.0
sniffio==1.3.1
SQLAlchemy==2.0.40
stack-data==0.6.3
sympy==1.14.0
tenacity==8.5.0
textual==5.0.1
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
triton==3.3.0
typing_extensions==4.13.2
typing-inspect==0.9.0
typing-inspection==0.4.0
uc-micro-py==1.0.3
urllib3==2.2.3
uvicorn==0.24.0
wcwidth==0.2.13
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
firebase-admin==6.5.0
email-validator==2.1.1
bcrypt==4.1.2
